+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add `--tag <names-or-ids>` option to `task add` to attach tags (by name or ID) at creation time, instead of requiring a separate `tag attach` command (#650)
+
 ## [3.16.0] - 2026-07-04
 
 ### Added